requirements.txt
biopython>=1.83
openpyxl>=3.1
matplotlib>=3.8
